Organize logic (#2358)
* clean up reused function

* remove fluff and use constants

* standardizing compare

* segment logic for clickhouse queries that have a lot

* add default to filter.date_to

* mino unwrap

* fix test
