+++ ee/clickhouse/queries/clickhouse_paths.py
@@ -1,4 +1,4 @@
-from typing import Any, Dict, List, Optional
+from typing import Dict, List, Optional
 
 from django.utils import timezone
 
@@ -10,11 +10,11 @@ from ee.clickhouse.sql.paths.path import PATHS_QUERY_FINAL
 from posthog.constants import AUTOCAPTURE_EVENT, CUSTOM_EVENT, SCREEN_EVENT
 from posthog.models.filter import Filter
 from posthog.models.team import Team
-from posthog.queries.base import BaseQuery
+from posthog.queries.paths import Paths
 from posthog.utils import relative_date_parse
 
 
-class ClickhousePaths(BaseQuery):
+class ClickhousePaths(Paths):
     def _determine_path_type(self, requested_type=None):
         # Default
         event: Optional[str] = "$pageview"
@@ -69,7 +69,7 @@ class ClickhousePaths(BaseQuery):
             path_type=path_type,
             parsed_date_from=parsed_date_from,
             parsed_date_to=parsed_date_to,
-            filters=prop_filters if filter.properties else "",
+            filters=prop_filters,
             marked_session_start="{} = %(start_point)s".format(start_comparator)
             if filter and filter.start_point
             else "new_session",
@@ -96,6 +96,3 @@ class ClickhousePaths(BaseQuery):
 
         resp = sorted(resp, key=lambda x: x["value"], reverse=True)
         return resp
-
-    def run(self, filter: Filter, team: Team, *args, **kwargs) -> List[Dict[str, Any]]:
-        return self.calculate_paths(filter=filter, team=team)
