+++ posthog/models/filter.py
@@ -182,13 +182,13 @@ class Filter(PropertyMixin):
         return timezone.now().replace(hour=0, minute=0, second=0, microsecond=0) - relativedelta(days=7)
 
     @property
-    def date_to(self) -> Optional[datetime.datetime]:
+    def date_to(self) -> datetime.datetime:
         if self._date_to:
             if isinstance(self._date_to, str):
                 return relative_date_parse(self._date_to)
             else:
                 return self._date_to
-        return None
+        return timezone.now()
 
     @property
     def date_filter_Q(self) -> Q:
