+++ posthog/test/test_filter_model.py
@@ -4,6 +4,7 @@ from typing import Any, Callable, Optional
 from dateutil.relativedelta import relativedelta
 from django.db.models import Q
 from django.utils import timezone
+from freezegun.api import freeze_time
 
 from posthog.api.test.base import BaseTest
 from posthog.models import Cohort, Element, Event, Filter, Person
@@ -357,17 +358,19 @@ class TestDateFilterQ(BaseTest):
         self.assertEqual(date_filter_query, Q())
 
     def test_default_filter_by_date_from(self):
-        filter = Filter(
-            data={
-                "properties": [
-                    {
-                        "key": "name",
-                        "value": json.dumps({"first_name": "Mary", "last_name": "Smith"}),
-                        "type": "person",
-                    }
-                ],
-            }
-        )
-        one_week_ago = timezone.now().replace(hour=0, minute=0, second=0, microsecond=0) - relativedelta(days=7)
-        date_filter_query = filter.date_filter_Q
-        self.assertEqual(date_filter_query, Q(timestamp__gte=one_week_ago))
+
+        with freeze_time("2020-01-01T00:00:00Z"):
+            filter = Filter(
+                data={
+                    "properties": [
+                        {
+                            "key": "name",
+                            "value": json.dumps({"first_name": "Mary", "last_name": "Smith"}),
+                            "type": "person",
+                        }
+                    ],
+                }
+            )
+            one_week_ago = timezone.now().replace(hour=0, minute=0, second=0, microsecond=0) - relativedelta(days=7)
+            date_filter_query = filter.date_filter_Q
+            self.assertEqual(date_filter_query, Q(timestamp__gte=one_week_ago, timestamp__lte=timezone.now()))
